Start splitting files apart for DOM rendering strategy
